CORDA-3584: Removed CordappResolver
